bb-downloader: Add integration tests
- Some unit tests were more like integration tests, so just move them to
  proper place.
- Allows testing public API properly.
